apps/agent/pyproject.toml
[project]
name = "orka-agent"
version = "0.1.0"
description = "orka-agent"
requires-python = ">=3.12.12"
dependencies = [
    "amdsmi>=7.0.2",
    "nvidia-ml-py>=13.595.45",
    "psutil>=7.2.2",
    "python-socketio[asyncio-client]>=5.16.1",
]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
